# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth service API documentation",
        "title": "api gat way",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/learning/ai/storage": {
            "get": {
                "description": "Get AI storage information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get storage from IA",
                "responses": {
                    "200": {
                        "description": "AI storage retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.AistorageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/challenge/solution/submit": {
            "post": {
                "description": "Submit a solution for a challenge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Submit solution for a challenge",
                "parameters": [
                    {
                        "description": "Submit Challenge Solution Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learning.SubmitChallengeSolutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solution submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.SubmitChallengeSolutionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/challenges": {
            "get": {
                "description": "Retrieves all challenges available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get all challenges",
                "responses": {
                    "200": {
                        "description": "Challenges retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.GetChallengesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/progress": {
            "get": {
                "description": "Retrieves progress details for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get progress for a user",
                "responses": {
                    "200": {
                        "description": "Progress details",
                        "schema": {
                            "$ref": "#/definitions/learning.ProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/quiz/submit": {
            "post": {
                "description": "Submit answers for a quiz",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Submit a quiz",
                "parameters": [
                    {
                        "description": "Submit Quiz Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learning.SubmitQuizRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.SubmitQuizResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/recommendations": {
            "get": {
                "description": "Retrieves topic recommendations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get recommendations",
                "responses": {
                    "200": {
                        "description": "Recommendations retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.GetRecommendationsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/resource/complete": {
            "post": {
                "description": "Mark a resource as completed for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Complete a resource",
                "parameters": [
                    {
                        "description": "Complete Resource Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learning.CompleteResourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource completed successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.CompleteResourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/resources": {
            "get": {
                "description": "Retrieves all resources available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get all resources",
                "responses": {
                    "200": {
                        "description": "Resources retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.GetResourcesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/topic/complete": {
            "post": {
                "description": "Mark a topic as completed for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Complete a topic",
                "parameters": [
                    {
                        "description": "Complete Topic Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learning.CompleteTopicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topic completed successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.CompleteTopicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Topic not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/topic/{topic_id}": {
            "get": {
                "description": "Get details of a specific topic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get a topic by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic ID",
                        "name": "topic_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topic details",
                        "schema": {
                            "$ref": "#/definitions/learning.Topic"
                        }
                    },
                    "404": {
                        "description": "Topic not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/topics": {
            "get": {
                "description": "Retrieves all topics available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get all topics",
                "responses": {
                    "200": {
                        "description": "Topics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/learning.GetTopicsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "learning.AistorageRespons": {
            "type": "object",
            "properties": {
                "request_text": {
                    "type": "string"
                },
                "response_text": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "learning.AistorageResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.AistorageRespons"
                    }
                }
            }
        },
        "learning.Challenge": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "xp_reward": {
                    "type": "integer"
                }
            }
        },
        "learning.CompleteResourceRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "learning.CompleteResourceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
        "learning.CompleteTopicRequest": {
            "type": "object",
            "properties": {
                "topic_id": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "learning.CompleteTopicResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
        "learning.Content": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "learning.GetChallengesResponse": {
            "type": "object",
            "properties": {
                "challenges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Challenge"
                    }
                }
            }
        },
        "learning.GetRecommendationsResponse": {
            "type": "object",
            "properties": {
                "recommendations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Topics"
                    }
                }
            }
        },
        "learning.GetResourcesResponse": {
            "type": "object",
            "properties": {
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Resource"
                    }
                }
            }
        },
        "learning.GetTopicsResponse": {
            "type": "object",
            "properties": {
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Topic"
                    }
                }
            }
        },
        "learning.ProgressResponse": {
            "type": "object",
            "properties": {
                "overall_progress": {
                    "type": "number"
                },
                "quizzes_completed": {
                    "type": "integer"
                },
                "resources_completed": {
                    "type": "integer"
                },
                "topics_completed": {
                    "type": "integer"
                },
                "total_quizzes": {
                    "type": "integer"
                },
                "total_resources": {
                    "type": "integer"
                },
                "total_topics": {
                    "type": "integer"
                }
            }
        },
        "learning.Quiz": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.QuizQuestion"
                    }
                },
                "topic_id": {
                    "type": "string"
                }
            }
        },
        "learning.QuizAnswer": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "string"
                },
                "selected_option": {
                    "type": "integer"
                }
            }
        },
        "learning.QuizQuestion": {
            "type": "object",
            "properties": {
                "correct_option": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "learning.Resource": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/learning.ResourceType"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "learning.ResourceType": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "ResourceType_ARTICLE",
                "ResourceType_INTERACTIVE"
            ]
        },
        "learning.SubmitChallengeSolutionRequest": {
            "type": "object",
            "properties": {
                "solution": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "learning.SubmitChallengeSolutionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "learning.SubmitQuizRequest": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.QuizAnswer"
                    }
                },
                "quiz_id": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "learning.SubmitQuizResponse": {
            "type": "object",
            "properties": {
                "correct_answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.QuizAnswer"
                    }
                },
                "feedback": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "total_questions": {
                    "type": "integer"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
        "learning.Topic": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Content"
                    }
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quiz": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/learning.Quiz"
                    }
                },
                "xpearned": {
                    "type": "integer"
                }
            }
        },
        "learning.Topics": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}